tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/015.html

<!doctype html>
<title> Link with javascript onclick and href script order </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="parent.events.push('click');" href="javascript:parent.events.push('href')">Test</a>
<script>
var t = async_test();
var events = [];
t.step(function() {
  document.getElementsByTagName("a")[0].click();
  events.push('after script');
});
onload = t.step_func(function() {
  // javascript: executions are async.
  assert_array_equals(events, ['click', 'after script', 'href']);
  t.done();
});
</script>
